Import new tool

+++ src/nanoprintf/nanoprintf.h
@@ -29,11 +29,6 @@
 extern "C" {
 #endif
 
-// The npf_ functions all return the number of bytes required to express the
-// fully-formatted string, not including the null terminator character.
-// The npf_ functions do not return negative values, since the lack of 'l' length
-// modifier support makes encoding errors impossible.
-
 NPF_VISIBILITY int npf_snprintf(
   char *buffer, size_t bufsz, const char *format, ...) NPF_PRINTF_ATTR(3, 4);
 
@@ -517,7 +512,7 @@ int npf_fsplit_abs(float f, uint64_t *out_int_part, uint64_t *out_frac_part,
 
   if (exponent >= (64 - NPF_MANTISSA_BITS)) { return 0; } // value is out of range
 
-  uint32_t const implicit_one = ((uint32_t)1) << NPF_MANTISSA_BITS;
+  uint32_t const implicit_one = 1u << NPF_MANTISSA_BITS;
   uint32_t const mantissa = f_bits & (implicit_one - 1);
   uint32_t const mantissa_norm = mantissa | implicit_one;
